requirements.txt
# Sistema Inteligente para Otimização de Pulverização de Agrotóxicos
# TCC - ETEC Rodrigues de Abreu

############################################################
# Núcleo NumPy (compatibilidade com OpenCV / scikit-learn)
############################################################
numpy>=1.24.0,<2.0

############################################################
# Machine Learning (leve)
############################################################
scikit-learn==1.3.0
pandas==2.0.3
joblib==1.3.2

############################################################
# Visão Computacional
############################################################
opencv-python-headless>=4.8.0,<5.0.0
Pillow>=10.0.0

############################################################
# API / Web
############################################################
fastapi==0.103.1
starlette==0.27.0
uvicorn==0.23.2
pydantic>=2.3.0,<3.0.0
pydantic-settings>=2.0.0
python-multipart==0.0.6

############################################################
# Banco de Dados
############################################################
SQLAlchemy>=2.0.0,<2.1.0
psycopg2-binary>=2.9.9

############################################################
# Integrações / Utilidades
############################################################
requests==2.31.0
aiohttp==3.8.5
httpx==0.24.1
python-dotenv>=1.0.0

############################################################
# Logging / Monitoramento
############################################################
loguru>=0.7.0
prometheus-client>=0.17.0

############################################################
# Segurança
############################################################
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

############################################################
# Utilitários
############################################################
tqdm>=4.66.0
python-dateutil>=2.8.2
pytz>=2023.3
colorama>=0.4.6

############################################################
# Desenvolvimento (opcional)
############################################################
pytest>=7.4.0
black>=23.9.1
flake8>=6.1.0
